Improve softbody grasping detection and grasping itself. Add ability to easily draw. StochasticPhysicsTracker.
Changes
PR2SoftBodyGripper::attach() sets anchors at the transform of the middle of the
fingers, instead of a particular finger.
PR2SoftBodyGripper::attach() projects the nodes to the plane at the middle of
the fingers before appending the anchor to that node.
In PR2SoftBodyGripper::attach(), a node is not in a grasp region if the node is
farther by some distance from the plane at the middle of the fingers; before,
a node is not in a grasp region if the node is not between the fingers.
Changed initialization parameters for cloth with intuitive ones.

New features
Draw once in the scene, openGL style. Useful for drawing transforms, lines,
points, etc for debugging.
Grab detector (HysteresisGrabDetector) that grabs and releases based on the
gripper angle and hystersis; detection is more robust than the GrabDetector.
StochasticPhysicsTracker