Add HumanHandObject and fix center of mass offsets.
Changes
Fix center of mass offset for meshes (before, the center of mass could be
outside the object and thus give an unrealistic physical behavior).
Fix graphicsMesh option, which indicates if the exact mesh or the approximated
mesh (if applicable) should be rendered.

New features
Add human hand model file and a human hand class.
Add the PUGIXML library, an XML parser. It's currently not being used.